Adds functionality for forgot and update password If someone forgets their password they will be sent a temporary password and that lasts for 10 minutes. Their account is marked as using a temp password on login the frontend should prompt to make a new password when this flag is present. Changes also include endpoint to be able to make said update to password.

+++ back-end/access-control/src/main/java/com/appdomain/accesscontrol/security/services/UserService.java
@@ -1,12 +1,14 @@
 package com.appdomain.accesscontrol.security.services;
 
 import com.appdomain.accesscontrol.security.contracts.CustomUserDetails;
+import com.appdomain.accesscontrol.security.contracts.PasswordUpdateRequest;
 import com.appdomain.accesscontrol.security.contracts.UserRegistrationRequest;
 import com.appdomain.accesscontrol.security.domains.User;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.HttpClientErrorException;
 
@@ -15,14 +17,18 @@ import java.util.Calendar;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.appdomain.accesscontrol.security.utils.ConstraintUtils.isValidPassword;
+
 @Service
 public class UserService {
 
     private final CustomUserDetailsService userDetailsService;
+    private final PasswordEncoder passwordEncoder;
 
     @Autowired
-    public UserService(CustomUserDetailsService userDetailsService) {
+    public UserService(final CustomUserDetailsService userDetailsService, final PasswordEncoder passwordEncoder) {
         this.userDetailsService = userDetailsService;
+        this.passwordEncoder = passwordEncoder;
     }
 
     public void processRegistration(final UserRegistrationRequest registrationRequest) {
@@ -33,24 +39,30 @@ public class UserService {
                     registrationRequest.getUserEmail());
         }
 
-        final CustomUserDetails currentUser;
-        try {
-            currentUser = (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        } catch (Exception e) {
-            throw HttpClientErrorException.create("User Context not found", HttpStatus.UNAUTHORIZED,
-                    "", null,null,null);
+        final CustomUserDetails currentAdmin = this.getCurrentUser();
+        if (!currentAdmin.getUser().getRole().equals("ROLE_ADMIN")) {
+            throw HttpClientErrorException.create("Current user is not authorized for this action",
+                    HttpStatus.UNAUTHORIZED, "",null,null,null);
         }
-
         if (registrationRequest.isApproved()) {
             user.setUserName(getUserName(user));
             user.setRole(registrationRequest.getRole());
-            user.setRegisteredBy(currentUser.getUser().getId());
+            user.setRegisteredBy(currentAdmin.getUser().getId());
             user.setLockoutEnd(Instant.now());
             user.setAwaitingRegistration(false);
             //TODO: Send email to user stating that their account has been confirmed
         } else {
             this.userDetailsService.deleteUser(user);
             //TODO: Send email to the user including registrationRequest.denialReason();
+        }
+    }
+
+    private CustomUserDetails getCurrentUser() {
+        try {
+            return (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        } catch (Exception e) {
+            throw HttpClientErrorException.create("User Context not found", HttpStatus.UNAUTHORIZED,
+                    "", null,null,null);
         }
     }
 
@@ -69,5 +81,24 @@ public class UserService {
 
     public List<CustomUserDetails> getUnregisteredUsers() {
         return this.userDetailsService.loadAllUsersAwaitingRegistration();
+    }
+
+
+    public void updatePassword(final PasswordUpdateRequest updateRequest) {
+        if (!isValidPassword(updateRequest.getNewPass())) {
+            //TODO make special exception (see LoginService.java)
+            throw HttpClientErrorException.create("Password does not meet rules", HttpStatus.BAD_REQUEST,
+                    "",null,null,null);
+        }
+
+        final User user = this.getCurrentUser().getUser();
+        if (!user.isTempPassword()) {
+            //TODO: Register users old password hash in password history table
+        }
+
+        user.setTempPassword(false);
+        user.setPasswordHash(this.passwordEncoder.encode(updateRequest.getNewPass()));
+        user.setPasswordCreateDate(Instant.now());
+        this.userDetailsService.saveCustomUser(user);
     }
 }
